remove decimel

+++ database/migrations/2023_12_28_022549_create_exchange_rates_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('exchange_rates', function (Blueprint $table) {
             $table->id();
-            $table->decimal('jpy', 20, 4);
-            $table->decimal('usd', 20,4);
+            $table->float('jpy');
+            $table->float('usd');
             $table->string('exchange_rate_month', 100)->nullable();
             $table->timestamps();
             $table->softDeletes();
