+++ database/migrations/2023_12_28_022717_create_outsourcings_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('user_id');
             $table->string('company_name', 255);
-            $table->decimal('jpy', 20, 4)->nullable();
-            $table->decimal('usd', 20, 4)->nullable();
-            $table->decimal('vnd', 20, 4)->nullable();
+            $table->float('jpy')->nullable();
+            $table->float('usd')->nullable();
+            $table->float('vnd')->nullable();
             $table->unsignedBigInteger('exchange_rate_id');
             $table->string('outsourced_project', 255)->nullable();
             $table->dateTime('outsourced_date')->nullable();
