+++ database/migrations/2024_01_04_082727_create_balance_orders_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('balance_sheets', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->decimal('amount', 20, 4);
+            $table->float('amount');
             $table->string('bs_month_year', 50);
             $table->unsignedBigInteger('category_id');
             $table->timestamps();
